# docs/swagger.yaml
definitions:
  response.Resp:
    properties:
      code:
        $ref: '#/definitions/xcode.Xcode'
      data: {}
      msg:
        type: string
      timestamp:
        type: integer
    type: object
  v1.LoginReq:
    properties:
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  v1.LogoutReq:
    properties:
      refreshToken:
        description: 刷新令牌
        type: string
    type: object
  v1.RefreshReq:
    properties:
      refreshToken:
        description: 刷新令牌
        type: string
    type: object
  v1.UserCreateReq:
    properties:
      avatar:
        description: 头像地址
        type: string
      email:
        description: 邮箱
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      username:
        description: 用户名
        type: string
    type: object
  xcode.Xcode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2005
    - 2006
    - 3000
    - 3001
    - 3002
    - 3003
    - 3004
    - 4000
    - 4001
    - 4002
    - 4003
    - 4004
    - 4005
    - 4006
    - 4007
    - 4008
    - 4009
    format: int32
    type: integer
    x-enum-comments:
      CacheError: Cache Error
      CreateSuccess: Create Success
      DatabaseError: Database Error
      DeleteSuccess: Delete Success
      ErrInvalidParam: Invalid Parameter
      ExternalAPIFail: External API Failure
      FileTypeInvalid: Invalid File Type
      FileUploadFail: File Upload Failed
      Forbidden: Forbidden
      LoginFailed: Login Failed
      MethodNotAllow: Method Not Allowed
      MissingParam: Missing Parameter
      NotFound: Not Found
      ParamError: Parameter Error
      PasswordError: Password Error
      PermissionAlreadyExist: Permission Already Exists
      PermissionDenied: Permission Denied
      ServerError: Internal Server Error
      Success: Success
      TimeoutError: Timeout
      TokenExpired: Token Expired
      TokenInvalid: Token Invalid
      Unauthorized: Unauthorized
      UpdateSuccess: Update Success
      UserAlreadyExist: User Already Exists
      UserNotExist: User Not Exists
    x-enum-descriptions:
    - Success
    - Create Success
    - Delete Success
    - Update Success
    - Parameter Error
    - Missing Parameter
    - Method Not Allowed
    - Unauthorized
    - Forbidden
    - Not Found
    - Invalid Parameter
    - Internal Server Error
    - Database Error
    - Cache Error
    - External API Failure
    - Timeout
    - File Upload Failed
    - Invalid File Type
    - User Already Exists
    - User Not Exists
    - Password Error
    - Token Expired
    - Token Invalid
    - Permission Denied
    - Permission Already Exists
    - Login Failed
    x-enum-varnames:
    - Success
    - CreateSuccess
    - DeleteSuccess
    - UpdateSuccess
    - ParamError
    - MissingParam
    - MethodNotAllow
    - Unauthorized
    - Forbidden
    - NotFound
    - ErrInvalidParam
    - ServerError
    - DatabaseError
    - CacheError
    - ExternalAPIFail
    - TimeoutError
    - FileUploadFail
    - FileTypeInvalid
    - UserAlreadyExist
    - UserNotExist
    - PasswordError
    - TokenExpired
    - TokenInvalid
    - PermissionDenied
    - PermissionAlreadyExist
    - LoginFailed
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口
      parameters:
      - description: 登录请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/v1.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/response.Resp'
      summary: 用户登录
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 用户登出接口
      parameters:
      - description: 登出请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/v1.LogoutReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            $ref: '#/definitions/response.Resp'
      summary: 用户登出
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 刷新Token接口
      parameters:
      - description: 刷新Token请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/v1.RefreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            $ref: '#/definitions/response.Resp'
      summary: 刷新Token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 用户注册接口
      parameters:
      - description: 注册请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/v1.UserCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/response.Resp'
      summary: 用户注册
      tags:
      - Auth
swagger: "2.0"